bb11742 fix compile error in bytecode_api.c on Mac OS X.

+++ libclamav/bytecode_api.c
@@ -123,8 +123,8 @@ int32_t cli_bcapi_seek(struct cli_bc_ctx* ctx, int32_t pos, uint32_t whence)
             return -1;
     }
     if (off < 0 || off > ctx->file_size) {
-        cli_dbgmsg("bcapi_seek: out of file: %ld (max %d)\n",
-                   off, ctx->file_size);
+        cli_dbgmsg("bcapi_seek: out of file: %lld (max %d)\n",
+                   (long long)off, ctx->file_size);
         return -1;
     }
     cli_event_int(EV, BCEV_OFFSET, off);
